Move egg functions around
